requirements.txt
python-telegram-bot==5.3.0
requests==2.12.4